Hard panic in task runs; scheduler timeout for pipeline run
Recovering panics in the pipeline can lead to inconsistent state and
hang the pipeline indefinitely. It is better to hard panic immediately
and allow the node to be cleanly restarted.

In addition, the pipeline runner should have some kind of sanity
timeout. This commit re-adds a deadline for JobPipelineMaxRunDuration
which appears to have been removed by accident at some point.