Add an interface to replace the ChangeSet function type. (#16186)
This preserves the existing function type (as an alias to a new function type, ChangeLogic).  It adds also two function types ChangeLogic and PreconditionVerifier, to encapsulate the logic of these two functions. The interface exposes an Apply and a VerifyPreconditions method. The aforementioned function types can be used with a convenience function to generate an implementation of ChangeSetV2. Alternatively, a full implementation of the interface can be built using a struct that implements those methods.

Also:
  * Rework ChangesetApplication to pass around an interface ConfiguredChangeSet, which can be generic, and therefore avoid a lot of mess in dealing with the WrapChangeset stuff.
  * Pull out WrapChangeset, which existed to handle some generics which are now handled by the ConfiguredChangeSet interface.
  * Wire in validation to the execution of changesets in the test infrastructure, and add some tests to make sure that works.